% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\name{aapply}
\alias{aapply}
\title{aapply}
\usage{
aapply(SE, fx, assay_type = 1)
}
\arguments{
\item{function:}{function to apply on the nested dataframes}

\item{assay_type:}{integer or name of the assay on which to apply the function}
}
\value{
the same SE object with updated nested dataframe
}
\description{
works like sapply but on each nested dataframe of the assay of an SE
}
