% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{add_Drug_annotation}
\alias{add_Drug_annotation}
\title{add_Drug_annotation}
\usage{
add_Drug_annotation(df_metadata)
}
\arguments{
\item{df_metadata}{a dataframe with metadata}
}
\value{
a dataframe with metadata with annotated drugs
}
\description{
add drug annotation to a dataframe with metadata
}
\examples{
add_Drug_annotation(df_metadata)
}
