% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\name{addAssayToMAE}
\alias{addAssayToMAE}
\title{addAssayToMAE}
\usage{
addAssayToMAE(mae, assay, assay_name, exp_name = c("treated",
  "untreated"), update_assay = FALSE)
}
\arguments{
\item{mae}{MultiAssayExperiment object with dose-reponse}

\item{assay}{matrix with dose-response data}

\item{assay_name}{string name of the assay to be used in SE}

\item{exp_name}{string name of the MAE experiment (i.e. SEs) to which add the assay}

\item{update_assay}{logical allow for assay update if the assay with 'assay_name' currently exists in given 'exp_name'}
}
\value{
MultiAssayExperiment object with dose-reponse data
}
\description{
Add assay to one of MAEs experiments (i.e. SEs) with dose-reponse data
}
