% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R, R/ttt.R
\name{assay_to_df}
\alias{assay_to_df}
\title{assay_to_df}
\usage{
assay_to_df(se, assay_name)

assay_to_df(se, assay_name)
}
\arguments{
\item{se}{SummarizedExperiment object with dose-response data}

\item{assay_name}{string name of the assay}

\item{se}{SummarizedExperiment object with dose-response data}

\item{assay_name}{string name of the assay}
}
\value{
data.frame with dose-reponse data

data.frame with dose-reponse data
}
\description{
Convert SE assay to data.frame

Convert SE assay to data.frame
}
