% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{average_replicates}
\alias{average_replicates}
\title{average_replicates}
\usage{
average_replicates(df_normalized, TrtKeys = NULL)
}
\arguments{
\item{df_normalized}{a dataframe with normalized values of DR data}

\item{TrtKeys}{a vector of keys used for averaging (NULL by default)}
}
\value{
a dataframe with averaged replicated of DR data
}
\description{
average_replicates
}
\examples{
average_replicates(df_normalized)
}
