% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_SE.R
\name{average_SE}
\alias{average_SE}
\title{average_SE}
\usage{
average_SE(
  se,
  TrtKeys = NULL,
  include_masked = FALSE,
  aggregate_FXN = function(x) {     mean(x, na.rm = TRUE, trim = 0.25) },
  treated_assay = "RawTreated",
  reference_assay = "UntreatedReferences"
)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} with drug response data.}

\item{TrtKeys}{a vector of keys used for averaging (NULL by default)}

\item{aggregate_FXN}{function used for averaging data that should be considered replicates.
Defaults to trimmed arithmetic mean with trim = 0.25.}
}
\value{
a SummarizedExperiment with additional assay with averaged DR data
}
\description{
Average the assays of a SummarizedExperiment of drug response data
}
