% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calculate_DRmetrics}
\alias{calculate_DRmetrics}
\title{calculate_DRmetrics}
\usage{
calculate_DRmetrics(df_averaged, DoseRespKeys = NULL,
  studyConcThresh = 4)
}
\arguments{
\item{df_averaged}{a SummarizedExperiment with averaged and normalized assays}

\item{DoseRespKeys}{a vector of dose response keys used for calculation of metrics (NULL by default)}

\item{studyConcThresh}{a numeric with study concentration threshold (4 by default)}
}
\value{
a dataframe with with metrics
}
\description{
Calculate metrics for DR data
}
\examples{
\dontrun{
calculate_DRmetrics(df_averaged)
}
}
