% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_GR_value.R
\name{calculate_GR_value}
\alias{calculate_GR_value}
\title{Calculate a GR value.}
\usage{
calculate_GR_value(
  rel_viability,
  corrected_readout,
  day0_readout,
  untrt_readout,
  ndigit_rounding,
  duration,
  ref_div_time,
  cl_name = "cell line"
)
}
\arguments{
\item{rel_viability}{numeric vector representing the Relative Viability.}

\item{corrected_readout}{numeric vector containing the corrected readout.}

\item{day0_readout}{numeric vector containing the day 0 readout.}

\item{untrt_readout}{numeric vector containing the untreated readout.}

\item{ndigit_rounding}{integer specifying the number of digits to use for calculation rounding.}

\item{duration}{numeric value specifying the length of time the cells were treated.}

\item{ref_div_time}{numeric value specifying the reference division time for the cell line of interest.}

\item{cl_name}{character string specifying the name for the cell line of interest.}
}
\value{
numeric GR value.
}
\description{
Calculate a GR value for a given set of dose response values.
}
