% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyte_utils.R
\name{cast_to_wide}
\alias{cast_to_wide}
\title{Reshape a data.table from long to wide format}
\usage{
cast_to_wide(long_data, id_vars, col_var, val_var)
}
\arguments{
\item{long_data}{data.table; A data.table in long format.}

\item{id_vars}{character vector; Column names that uniquely identify each row (e.g. c("CellLineName", "range")).}

\item{col_var}{character; The column whose unique values will become the new column headers (e.g., "treatment").}

\item{val_var}{character; The column whose values will fill the cells (e.g., "rate").}
}
\value{
A new data.table in wide format.
}
\description{
This function utilizes the powerful dcast function from the data.table package
to pivot data. It turns unique values from a specified column into new
columns in the output data.table.
}
\keyword{incucyte}
