% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SE.R
\name{create_SE}
\alias{create_SE}
\title{create_SE}
\usage{
create_SE(
  df_data,
  data_type = c("untreated", "treated", "all"),
  readout = "ReadoutValue",
  discard_keys = NULL,
  assay_type = c("matrix", "BumpyMatrix"),
  aggregate_ref_FXN = function(x) {     mean(x, trim = 0.25) }
)
}
\arguments{
\item{df_data}{a dataframe with DR data}

\item{data_type}{string type of data to be returned: with untreated conditions only ('untreated'),
with treated conditions only ('treated') or all}

\item{readout}{a character that indiciated readout value in a df ('ReadoutValue' by default)}

\item{discard_keys}{a vector of keys that should be discarded}

\item{assay_type}{string assay type for the returned SummarizedExperiment}
}
\value{
SummarizedExperiment object with dose-reponse data
}
\description{
Create SummarizedExperiment object from data.frame(s) with dose-reponse data
}
