% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{create_symlink_to_current}
\alias{create_symlink_to_current}
\title{Create symbolic link to current version}
\usage{
create_symlink_to_current(base_dir, latest_version, verbose_mode = FALSE)
}
\arguments{
\item{base_dir}{Character. The base directory where the symbolic link should be created.}

\item{latest_version}{Character. The path to the latest version directory.}

\item{verbose_mode}{Logical, indicating if the run is for a verbose report.}
}
\description{
This function creates a symbolic link (\code{current} or \code{current_verbose})
that points to the latest version directory.
}
\examples{
# Assume "/path/to/dir/v3" is the latest version
\dontrun{
create_symlink_to_current("/path/to/dir", "/path/to/dir/v3")
}

}
\keyword{report}
