% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\name{createMAE}
\alias{createMAE}
\title{createMAE}
\usage{
createMAE(untreated, treated)
}
\arguments{
\item{untreated}{SummarizedExperiment object with dose-reponse data for 'untreated' conditions}

\item{treated}{SummarizedExperiment object with dose-reponse data for 'treated' conditions}
}
\value{
MultiAssayExperiment object with dose-reponse data
}
\description{
Create MultiAssayExperiment object from treated/untreated SEs with dose-reponse data
}
