% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\name{createSE}
\alias{createSE}
\title{createSE}
\usage{
createSE(
  df_data,
  data_type = c("untreated", "treated", "all"),
  readout = "ReadoutValue",
  discard_keys = NULL
)
}
\arguments{
\item{df_data}{a dataframe with DR data}

\item{data_type}{string type of data to be returned: with untreated conditions only ('untreated'),
with treated conditions only ('treated') or all}

\item{readout}{a character that indiciated readout value in a df ('ReadoutValue' by default)}

\item{discard_keys}{a vector of keys that should be discarded}
}
\value{
SummarizedExperiment object with dose-reponse data
}
\description{
Create SummarizedExperiment object from data.frame(s) with dose-reponse data
}
