% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\name{df_to_assay}
\alias{df_to_assay}
\title{df_to_assay}
\usage{
df_to_assay(
  data,
  data_type = c("all", "treated", "untreated"),
  discard_keys = NULL
)
}
\arguments{
\item{data}{tibble or data.frame with drug-response data}

\item{data_type}{string type of data to be returned: all, for untreated conditions only ('untreated')
or for treated conditions only ('treated')}

\item{discard_keys}{a vector of keys that should be discarded}
}
\value{
matrix
}
\description{
Convert data.frame with dose-reponse data to the assay.
}
\details{
The 'assay' object is simply a matrix with rownames being the treatment ids,
colnames being the ids of the cell lines and values being the DataFrames with
dose-response data for given cell lines under given conditions.
}
