% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SE_metadata.R
\name{.get_SE_metadata}
\alias{.get_SE_metadata}
\title{The primary purpose of this function is to allow other functions to create exposed getter functions.}
\usage{
.get_SE_metadata(se, name, subname = NULL)
}
\description{
The primary purpose of this function is to allow other functions to create exposed getter functions.
}
