% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{.guess_specs_of_delim_file}
\alias{.guess_specs_of_delim_file}
\title{return some specs regarding delimited files
currently:
- delimiter detection (e.g. csv, tsv)
- number of fields/columns (based on the line with the higher number of fields)}
\usage{
.guess_specs_of_delim_file(file, nrows = 10000, seps = c(",", "\\t"))
}
\arguments{
\item{file}{string}

\item{nrows}{integer number of rows used to make predictions}

\item{seps}{character vector with delimiters/separators that will be evaluated}
}
\value{
list with specs
}
\description{
return some specs regarding delimited files
currently:
- delimiter detection (e.g. csv, tsv)
- number of fields/columns (based on the line with the higher number of fields)
}
