% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{extract_mySQL}
\alias{extract_mySQL}
\title{Extract data from the mySQL db}
\usage{
extract_mySQL(
  project_id,
  condition_metadata,
  condition_additional_treatment,
  condition_codrug,
  treatment_metadata,
  response_mean,
  response_metrics
)
}
\arguments{
\item{project_id}{a character with a project id}

\item{condition_metadata}{a dataframe with condition metadata}

\item{condition_additional_treatment}{a dataframe with additional treatments}

\item{condition_codrug}{a dataframe with condition codrugs}

\item{treatment_metadata}{a datafram with treatment metadata}

\item{response_mean}{a dataframe with response means}

\item{response_metrics}{a dataframe with response metrics}
}
\value{
a list with averaged values and metrics of DR data
}
\description{
Extract data from the mySQL db
}
