% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyte_utils.R
\name{filter_treatments}
\alias{filter_treatments}
\title{Filter Treatments from SummarizedExperiment}
\usage{
filter_treatments(se, sel_trt)
}
\arguments{
\item{se}{SummarizedExperiment; The object containing the assay data.}

\item{sel_trt}{character vector; A list of treatment names to filter and extract.}
}
\value{
A data.table containing columns: treatment, CellCount_norm, CellLineName, and Duration.
}
\description{
Filters a SummarizedExperiment object to extract normalized cell counts for specific treatments
and formats them into a long data.table for plotting.
}
\keyword{incucyte}
