% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_SE.R
\name{fit_SE}
\alias{fit_SE}
\title{fit_SE}
\usage{
fit_SE(
  se,
  fitting_ref_assay = "FittingReferences",
  averaged_treated_assay = "AveragedTreated",
  metrics_assay = "Metrics",
  studyConcThresh = 4
)
}
\arguments{
\item{se}{a SummarizedExperiment with averaged and normalized assays}

\item{studyConcThresh}{a numeric with study concentration threshold (4 by default)}
}
\value{
a \code{SummarizedExperiment} with additional assay with metrics
}
\description{
Calculate metrics for DR data
}
