% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{format_mySQL}
\alias{format_mySQL}
\title{Refactor dataframes to be ready to add to the mySQL db}
\usage{
format_mySQL(
  df_averaged,
  df_metrics,
  project_id,
  condition_metadata_table_length,
  treatment_metadata_table_length
)
}
\arguments{
\item{df_averaged}{a dataframe with averaged values of DR data}

\item{df_metrics}{a dataframe with metrics of DR data}

\item{project_id}{a character with a project id}

\item{condition_metadata_table_length}{an integer with condition table length}

\item{treatment_metadata_table_length}{an integer with treatment table length}
}
\value{
a list with transformed elements of DR data
}
\description{
Refactor dataframes to be ready to add to the mySQL db
}
