% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SE_metadata.R
\name{get_SE_experiment_metadata}
\alias{get_SE_experiment_metadata}
\title{Get metadata for experiment metadata in SummarizedExperiment object.}
\usage{
get_SE_experiment_metadata(se)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} object for which to get experiment metadata.}
}
\value{
experiment metadata.
}
\description{
Get metadata for experiment metadata in SummarizedExperiment object metadata.
}
