% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SE_metadata.R
\name{get_SE_keys}
\alias{get_SE_keys}
\title{Get metadata for keys in SummarizedExperiment object.}
\usage{
get_SE_keys(se, key_type = NULL)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} object for which to get metadata relating to keys.}

\item{key_type}{string of a specific}
}
\description{
Get metadata for keys in SummarizedExperiment object metadata.
}
