% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{getMetaData}
\alias{getMetaData}
\title{getMetaData}
\usage{
getMetaData(data, discard_keys = NULL)
}
\arguments{
\item{data}{data.frame with drug-response data}

\item{discard_keys}{character vector of keys to discard from the data columns}
}
\value{
named list containing different elements of a \linkS4class{SummarizedExperiment};
see details.
}
\description{
Get metadata of given project
}
\details{
Named list containing the following elements:
\itemize{
 \item{rowData}{treatment metadata}
 \item{colData}{condition metadata}
 \item{dataCols}{all data.frame columns corresponding to non-metadata fields}
 \item{csteData}{metadata that is constant for all entries of the data.frame}
}
}
