% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{ICGRfits}
\alias{ICGRfits}
\title{Actual fitting function}
\usage{
ICGRfits(df_, e_0 = 1, GR_0 = 1, force = FALSE)
}
\arguments{
\item{df_}{a data frame used for calculation of fit parameters}

\item{e_0}{= 1 by default}

\item{GR_0}{= 1 by default}

\item{force}{use signifcance or not}
}
\value{
vector of parameters
}
\description{
\code{ICGRfits} returns fit parameters
}
\details{
returns fit parameters
}
