% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{ICGRlogisticFit}
\alias{ICGRlogisticFit}
\title{Actual fitting function}
\usage{
ICGRlogisticFit(
  log10concs,
  RelativeViability,
  GRvalues,
  e_0 = 1,
  GR_0 = 1,
  force = FALSE,
  cap = FALSE
)
}
\arguments{
\item{log10concs}{concentrations}

\item{RelativeViability}{values}

\item{GRvalues}{values}

\item{e_0}{=1 by default}

\item{GR_0}{=1 by default}

\item{force}{use signifcance or not}

\item{cap}{enforce e_0 and GR_0}
}
\value{
vector of parameters
}
\description{
\code{ICGRlogisticFit} returns fit parameters
}
\details{
returns fit parameters
}
