% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_keys.R
\name{identify_keys}
\alias{identify_keys}
\title{identify_keys}
\usage{
identify_keys(obj)
}
\arguments{
\item{obj}{a dataframe or SummarizedExperiment with keys}
}
\value{
a list of keys
}
\description{
Identify keys in the DR data represented by dataframe or SummarizedExperiment objects
}
