% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_keys.R
\name{identify_keys2}
\alias{identify_keys2}
\title{identify_keys2}
\usage{
identify_keys2(
  obj,
  discard_keys = c("Barcode", gDRutils::get_identifier("masked_tag"))
)
}
\arguments{
\item{obj}{a data.frame to identify keys for.}

\item{discard_keys}{character vector of keys to exclude from the returned list. 
The keys discarded should be identical to the keys in the third
dimension of the SummarizedExperiment.
Defaults to the \code{"Barcode"} and the \code{masked} identifier.}
}
\value{
named list of key types and their corresponding key values.
}
\description{
Identify keys in the DR data represented by dataframe or SummarizedExperiment objects
}
\seealso{
map_df, create_SE2
}
