% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import raw data}
\usage{
import_data(
  manifest_file,
  template_file,
  results_file,
  instrument = "EnVision",
  cell_line_annotation = NULL,
  drug_annotation = NULL
)
}
\arguments{
\item{manifest_file}{character, file path(s) to manifest(s)}

\item{template_file}{character, with datapaths and names of results file(s)
or character with file path of templates file(s)}

\item{results_file}{character, with datapaths and names of results file(s)
or character with file path of results file(s)}

\item{instrument}{string with type of instrument used}

\item{cell_line_annotation}{optional data.table with cell line annotations;
if NULL (default), annotations are looked up from gDRinternal or gDRtestData}

\item{drug_annotation}{optional data.table with drug annotations;
if NULL (default), annotations are looked up from gDRinternal or gDRtestData}
}
\value{
a \code{data.table}
}
\description{
Import raw data
}
\examples{
td <- get_test_data()
i_df <- import_data(manifest_path(td), template_path(td), result_path(td))

}
\keyword{import}
