% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_utils.R
\name{is.readable.v}
\alias{is.readable.v}
\title{Check if all paths in vector are readable}
\usage{
is.readable.v(paths)
}
\arguments{
\item{paths}{a character with path(s)}
}
\value{

}
\description{
Check if all paths in vector are readable
}
