% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{is_null_or_empty}
\alias{is_null_or_empty}
\title{Check if an Object is NULL or Empty}
\usage{
is_null_or_empty(x)
}
\arguments{
\item{x}{An object to be checked. This can be any R object, but it is typically a vector or list.}
}
\value{
A logical value: \code{TRUE} if the object is \code{NULL} or contains any empty strings, \code{FALSE} otherwise.
}
\description{
This function checks whether a given object is either \code{NULL} or contains any empty strings.
}
\keyword{internal}
