% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_manifest}
\alias{load_manifest}
\title{Load manifest}
\usage{
load_manifest(manifest_file, log_str)
}
\arguments{
\item{manifest_file}{character, file path(s) to manifest(s)}

\item{log_str}{character, file path to logs}
}
\description{
This functions loads and checks the manifest file(s)
}
