% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_templates}
\alias{load_templates}
\title{Load templates}
\usage{
load_templates(df_template_files, log_str)
}
\arguments{
\item{df_template_files}{data.frame, with datapaths and names of results file(s)
or character with file path of templates file(s)}

\item{log_str}{character, file path to logs}
}
\description{
This functions loads and checks the template file(s)
}
