% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{logistic_4parameters}
\alias{logistic_4parameters}
\title{logistic function (not used in the file but useful for plotting externally)}
\usage{
logistic_4parameters(c, Vinf, V0, EC50, h)
}
\arguments{
\item{c}{c value}

\item{Vinf}{Vinf value}

\item{V0}{V0 value}

\item{EC50}{EC50 value}

\item{h}{h value}
}
\description{
logistic function (not used in the file but useful for plotting externally)
}
