% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{logisticFit}
\alias{logisticFit}
\title{Actual fitting function}
\usage{
logisticFit(concs, normValues, x_0 = 1, curve_type = c("IC", "GR"),
  force = FALSE, cap = 0.1)
}
\arguments{
\item{normValues}{values}

\item{x_0}{upper limit; =1 by default )}

\item{curve_type}{response curve: either IC ([0,1]) or GR([-1,1])}

\item{force}{use signifcance or not}

\item{cap}{enforce x_0}

\item{log10concs}{log10 of concentrations}
}
\value{
vector of parameters
}
\description{
\code{logisticFit} returns fit parameters
}
\details{
returns fit parameters
}
