% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df.R
\name{map_df}
\alias{map_df}
\title{Map treated conditions to their respective references.}
\usage{
map_df(
  trt_md,
  ref_md,
  row_endpoint_value_filter,
  Keys,
  ref_type = c("Day0", "untrt_Endpoint", "ref_Endpoint")
)
}
\arguments{
\item{trt_md}{data.frame of treated metadata.}

\item{ref_md}{data.frame of untreated metadata.}

\item{Keys}{named list of keys and values.
Likely obtained from \code{identify_keys()}.}

\item{ref_type}{string of the reference type to map to.
Should be one of \code{c("Day0", "untrt_Endpoint", "ref_Endpoint")}.}
}
\value{
named list mapping treated metadata to untreated metadata.
}
\description{
Map treated conditions to their respective Day0, untreated, or single-agent references using condition metadata.
}
\seealso{
identify_keys2
}
