% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{mapSE}
\alias{mapSE}
\title{mapSE}
\usage{
mapSE(normSE, ctrlSE, row_endpoint_value_filter, Keys, T0 = FALSE)
}
\arguments{
\item{normSE}{a SummarizedExperiment with normalization assaay}

\item{T0}{a logical indicating if the mapping should be performer for Time=0 (FALSE by default)}
}
\value{
a list of mapping
}
\description{
Perfmorm mapping for normalization
}
\examples{
\dontrun{
mapSE(normSE, ctrlSE, row_endpoint_value_filter)
}
}
