% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{merge_data}
\alias{merge_data}
\title{merge_data}
\usage{
merge_data(manifest, treatments, data)
}
\arguments{
\item{manifest}{a data frame with a manifest info}

\item{treatments}{a data frame with a treaatments info}

\item{data}{a data frame with a raw data info}
}
\value{
a dataframe with merged data
}
\description{
Merge all the input data into a single dataframe
}
\examples{
merge_data(manifest, treatments, data)
}
