% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{normalize_data}
\alias{normalize_data}
\title{normalize_data}
\usage{
normalize_data(df_raw_data, selected_keys = NULL, key_values = NULL)
}
\arguments{
\item{df_raw_data}{a dataframe with raw data}

\item{selected_keys}{a vector of keys that should be included in the normalization (NULL by default)}

\item{key_values}{a list of values for keys that should be included in the normalization (NULL by default)}
}
\value{
a dataframe with normalization
}
\description{
Normalize raw DR data
}
\examples{
normalize_data(df_raw_data)
}
