% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_SE.R
\name{normalize_SE}
\alias{normalize_SE}
\title{normalize_SE}
\usage{
normalize_SE(se, nDigits_rounding = 4)
}
\arguments{
\item{se}{\code{BumpyMatrix} object with assays \code{"treated"} and \code{Controls}.}

\item{nDigits_rounding}{integer specifying number of digits of rounding during calculations.
Defaults to \code{4}.}
}
\value{
\code{BumpyMatrix} object with a new assay named \code{"Normalized"} containing \code{DataFrame}s 
holding \code{RelativeViability} and \code{GRvalue} values.
}
\description{
Normalize drug response data from treated and untreated pairings.
}
