% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_SE.R
\name{normalize_SE2}
\alias{normalize_SE2}
\title{normalize_SE2}
\usage{
normalize_SE2(se, trt_keys = NULL, ndigit_rounding = 4)
}
\arguments{
\item{se}{\code{BumpyMatrix} object with assays \code{"RawTreated"} and \code{"Controls"}.}

\item{ndigit_rounding}{integer specifying number of digits of rounding during calculations.
Defaults to \code{4}.}
}
\value{
\code{BumpyMatrix} object with a new assay named \code{"Normalized"} containing \code{DataFrame}s 
holding \code{RelativeViability}, \code{GRvalue}, \code{RefRelativeViability}, code{RefGRvalue}, and \code{DivisionTime} values.
}
\description{
Normalize drug response data from treated and untreated pairings.
}
