% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_SE.R
\name{normalize_trt_to_ref}
\alias{normalize_trt_to_ref}
\title{Normalize a set of treated drug response values to its corresponding reference drug response values.}
\usage{
normalize_trt_to_ref(trt_df, ref_df, ndigits_rounding)
}
\arguments{
\item{trt_df}{data.frame of the treated elements.}

\item{ref_df}{data.frame of the reference elements to normalize against.}

\item{ndigits_rounding}{integer of the number of digits to round to during normalization calculations.}
}
\value{
DataFrame of containing the \code{RelativeViability}, \code{GRvalue}, adn \code{CorrectedReadout}.
}
\description{
Normalize treated and reference pairings to obtain key metrics on a per-concentration basis.
}
