% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Overall_function}
\alias{Overall_function}
\title{Deprecated function(s) in \pkg{gDR}
Overall_function}
\usage{
Overall_function(
  manifest_file,
  template_file,
  results_file,
  output_files,
  selected_keys = NULL,
  key_values = NULL,
  instrument = "EnVision"
)
}
\arguments{
\item{manifest_file}{a file path for the manifest file}

\item{template_file}{a file path(s) for the template file(s)}

\item{results_file}{a file path(s) for the results file(s)}

\item{output_files}{}

\item{selected_keys}{a vector of keys that should be included in the normalization}

\item{key_values}{a list of key values that should be included in the normalization}

\item{instrument}{character}
}
\value{
a list with the raw, normalized, averaged, and metrics dataframes
}
\description{
Process DR data using set of manifest, template and results files
}
