% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{range_mv}
\alias{range_mv}
\title{range_mv}
\usage{
range_mv(
  c50,
  x_inf,
  hillCoefficient,
  x_0 = 1,
  linear_conc_range_uM = c(0.005, 5)
)
}
\arguments{
\item{c50}{c50 values}

\item{x_inf}{x_inv values}

\item{hillCoefficient}{hill ceofficient}

\item{x_0}{= 1 by default}

\item{linear_conc_range_uM}{= c(5e-3, 5) by default}
}
\value{

}
\description{
range_mv
}
