% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRcomponents.R, R/gDRcore.R, R/gDRimport.R,
%   R/gDRutils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{getDrugs}
\alias{getCellLines}
\alias{getProjects}
\alias{getMetaProjects}
\alias{pullProjectSE}
\alias{merge_data}
\alias{runDrugResponseProcessingPipeline}
\alias{create_and_normalize_SE}
\alias{create_SE}
\alias{normalize_SE}
\alias{average_SE}
\alias{fit_SE}
\alias{load_data}
\alias{fit_curves}
\alias{logisticFit}
\alias{get_header}
\alias{reset_headers}
\alias{get_identifier}
\alias{set_identifier}
\alias{reset_identifiers}
\alias{convert_se_assay_to_dt}
\alias{validate_SE}
\alias{flatten}
\alias{split_SE_components}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gDRcomponents}{\code{\link[gDRcomponents:getResource]{getCellLines}}, \code{\link[gDRcomponents:getResource]{getDrugs}}, \code{\link[gDRcomponents:getResource]{getMetaProjects}}, \code{\link[gDRcomponents:getResource]{getProjects}}, \code{\link[gDRcomponents:getResource]{pullProjectSE}}}

  \item{gDRcore}{\code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{average_SE}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{create_SE}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{create_and_normalize_SE}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{fit_SE}}, \code{\link[gDRcore]{merge_data}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{normalize_SE}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{runDrugResponseProcessingPipeline}}}

  \item{gDRimport}{\code{\link[gDRimport]{load_data}}}

  \item{gDRutils}{\code{\link[gDRutils]{convert_se_assay_to_dt}}, \code{\link[gDRutils]{fit_curves}}, \code{\link[gDRutils]{flatten}}, \code{\link[gDRutils:headers]{get_header}}, \code{\link[gDRutils:identifiers]{get_identifier}}, \code{\link[gDRutils]{logisticFit}}, \code{\link[gDRutils:headers]{reset_headers}}, \code{\link[gDRutils:identifiers]{reset_identifiers}}, \code{\link[gDRutils:identifiers]{set_identifier}}, \code{\link[gDRutils]{split_SE_components}}, \code{\link[gDRutils]{validate_SE}}}
}}

