% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{run_gDRpipeline_with_report}
\alias{run_gDRpipeline_with_report}
\title{Run gDR pipeline with an accompanying report}
\usage{
run_gDRpipeline_with_report(
  qcs_id = NULL,
  manifest = NULL,
  treatment = NULL,
  raw_data = NULL,
  prism_data_path = NULL,
  cell_line_data_path = NULL,
  treatment_data_path = NULL,
  output_dir = getwd(),
  viz_format = "svg",
  push_to_DSDB = FALSE,
  rmd_template_path = system.file("report_templates", package = "gDR"),
  steps = seq_len(3),
  mae = NULL,
  special_sections = NULL
)
}
\arguments{
\item{qcs_id}{Character. The QCS ID. Provide this if you want to process data using a QCS ID.}

\item{manifest}{Character. Path to the manifest file. Required if not using \code{qcs_id} or \code{prism_data_path}.}

\item{treatment}{Character. Path to the treatment file. Required if not using \code{qcs_id} or \code{prism_data_path}.}

\item{raw_data}{Character. Path to the raw data file. Required if not using \code{qcs_id} or \code{prism_data_path}.}

\item{prism_data_path}{Character. Path to the PRISM data file. Provide this if you want to process PRISM data.}

\item{cell_line_data_path}{Character. Path to the cell line data file. Required if using \code{prism_data_path} and processing PRISM level 6 data.}

\item{treatment_data_path}{Character. Path to the treatment data file. Required if using \code{prism_data_path} and processing PRISM level 6 data.}

\item{output_dir}{Character. Directory to store reports and output data. This is a required parameter.}

\item{viz_format}{Character. Format for visualizations. Default is \code{"svg"}.}

\item{push_to_DSDB}{Logical. Whether to push results to DSDB. Default is \code{FALSE}.}

\item{rmd_template_path}{Character. Path to the RMarkdown template. Default is the directory of the current R script.}

\item{steps}{Integer vector. Steps to run in the pipeline (1: data import, 2: processing and QC, 3: analysis). Default is \code{1:3}. If \code{mae} is provided, only step 3 will be run.}

\item{mae}{MultiAssayExperiment or Character. A MultiAssayExperiment object or a path to a \code{.qs} file containing a MultiAssayExperiment object. If provided, only step 3 will be run and the MAE object will be saved to the \code{gDR_data} directory.}

\item{special_sections}{String or NULL. Provide this if you want to process data, type PRISM or chemical genomics, starting from \code{mae} to see a section specific to them. One of: "PRISM", "chemical_genomics".}
}
\value{
None. The function generates reports and saves them to the specified output directory.
}
\description{
This function runs the gDR pipeline and generates an accompanying report. The function can process different types of input data, including QCS ID, PRISM data, and a MultiAssayExperiment (MAE) object or a path to a \code{.qs} file containing an MAE object.
}
\examples{

\dontrun{
run_gDRpipeline_with_report("AS-42656", "~/test_report")
run_gDRpipeline_with_report(mae = "path/to/mae.qs", output_dir = "~/test_report")
}
}
