% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDrugResponseProcessingPipeline.R
\name{runDrugResponseProcessingPipeline2}
\alias{runDrugResponseProcessingPipeline2}
\title{Run drug response processing pipeline}
\usage{
runDrugResponseProcessingPipeline2(
  df_,
  readout = "ReadoutValue",
  control_mean_fxn = function(x) {     mean(x, trim = 0.25) },
  nested_keys = c("Barcode", gDRutils::get_identifier("masked_tag")),
  override_controls = NULL,
  include_masked = FALSE,
  ndigit_rounding = 4,
  control_assay = "Controls",
  raw_treated_assay = "RawTreated",
  normalized_assay = "Normalized",
  ref_GR_assay = "RefGRvalue",
  ref_RV_assay = "RefRelativeViability",
  averaged_assay = "Averaged",
  metrics_assay = "Metrics"
)
}
\arguments{
\item{df_}{data.frame of raw drug response data containing both treated and untreated values.}

\item{readout}{string of the name containing the cell viability readout values.}

\item{control_mean_fxn}{function indicating how to average controls.
Defaults to \code{mean(x, trim = 0.25)}.}

\item{nested_keys}{character vector of column names to include in the data.frames in the assays of the resulting \code{SummarizedExperiment} object.
Defaults to \code{c("Barcode", gDRutils::get_identifier("masked_tag"))}.}

\item{override_controls}{named list containing defining factors in the treatments.
Defaults to \code{NULL}.}

\item{include_masked}{boolean indicating whether or not to include masked wells
in the averaging.
This is used as an override to whatever wells have been masked in the original data.
Defaults to \code{FALSE}.}

\item{ndigit_rounding}{integer indicating number of digits to round to in calculations.
Defaults to \code{4}.}

\item{control_assay}{string containing the name of the assay representing the controls in the \code{se}.
Defaults to \code{"Controls"}.}

\item{raw_treated_assay}{string containing the name of the assay representing the raw treated data in the \code{se}.
Defaults to \code{"RawTreated"}.}

\item{normalized_assay}{string of the assay name containing the normalized data.
Defaults to \code{"Normalized"}.}

\item{ref_GR_assay}{string of the name of the reference GR assay in the \linkS4class{SummarizedExperiment}.
Defaults to \code{"RefGRvalue"}.}

\item{ref_RV_assay}{string of the name of the reference Relative Viability assay in the \linkS4class{SummarizedExperiment}.
Defaults to \code{"RefRelativeViability"}.}

\item{averaged_assay}{string of the name of the averaged assay in the \linkS4class{SummarizedExperiment}.
Defaults to \code{"Averaged"}.}

\item{metrics_assay}{string of the name of the metrics assay to output in the returned \linkS4class{SummarizedExperiment}
Defaults to \code{"Metrics"}.}
}
\description{
Create a SummarizedExperiment of raw data and proceed to 
normalize, average, and fit the processed data.
}
\seealso{
Other runDrugResponseProcessingPipelineFxns: 
\code{\link{average_SE2}()},
\code{\link{create_SE2}()},
\code{\link{fit_SE2}()},
\code{\link{normalize_SE2}()}
}
\concept{runDrugResponseProcessingPipelineFxns}
