% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SE_metadata.R
\name{set_SE_keys}
\alias{set_SE_keys}
\title{Set metadata for keys in SummarizedExperiment object.}
\usage{
set_SE_keys(se, value)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} object for which to add key metadata.}

\item{value}{named list of metadata for keys. 
Names of list should represent key types and list values should contain key type values.}
}
\value{
\linkS4class{se} with added metadata.
}
\description{
Set metadata for keys in SummarizedExperiment object metadata.
}
\seealso{
identify_keys2
}
