% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_data}
\alias{small_data}
\title{Small data.frame with raw data used for processing via gDR}
\format{
A data frame with 3300 rows and 12 variables:
\itemize{
\item{Barcode}
\item{clid}{cell line id}
\item{CellLineName}{cell line name}
\item{Tissue}{cell line tissue}
\item{ReferenceDivisionTime}{cell line division time}
\item{Gnumber}{drug id}
\item{DrugName}{drug name}
\item{drug_moa}{drug mode of action}
\item{Concentration}{drug concentration}
\item{ReadoutValue}{readout value}
\item{BackgroundValue}{backgroud value}
\item{Duration}{duration}
}
}
\usage{
data(small_data)
}
\value{
data.frame
}
\description{
A dataset containing the ReadoutValues for single-agent experiments
made-up of 10 drugs and 10 cell lines
}
\keyword{datasets}
