% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{split_comma_separated_string}
\alias{split_comma_separated_string}
\title{Split a comma-separated string into a character vector.}
\usage{
split_comma_separated_string(x)
}
\arguments{
\item{x}{A single character string or an existing character vector.}
}
\value{
A character vector with trimmed whitespace. If input is not a single string,
it is returned as is.
}
\description{
Split a comma-separated string into a character vector.
}
\keyword{internal}
