% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_SE_components.R
\name{split_SE_components}
\alias{split_SE_components}
\title{split_SE_components}
\usage{
split_SE_components(df_, discard_keys = NULL)
}
\arguments{
\item{df_}{data.frame with drug-response data}

\item{discard_keys}{character vector of keys to discard from the row or column data, 
and to leave in the matrix data. See details.}
}
\value{
named list containing different elements of a \linkS4class{SummarizedExperiment};
see details.
}
\description{
Divide the columns of an input data.frame into treatment metadata, condition metadata, 
experiment metadata, and core data for further analysis. This will most commonly be used
to identify the different components of a \linkS4class{SummarizedExperiment} object.
}
\details{
Named list containing the following elements:
\itemize{
 \item{treatment_md}{treatment metadata}
 \item{condition_md}{condition metadata}
 \item{data_fields}{all data.frame column names corresponding to fields represented within a BumpyMatrix cell}
 \item{experiment_md}{metadata that is constant for all entries of the data.frame}
}

The \code{discard_keys} provides the user the opportunity to specify that they would not 
like to use that metadata field as a differentiator of the treatments, and instead, incorporate it
into the \code{DataFrame} in the BumpyMatrix.

In the event that if any of the \code{discard_keys} are constant throughout the whole data.frame, 
they will still be included in the DataFrame of the BumpyMatrix and not in the experiment_metadata.
}
