% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utils.R
\name{test_lData}
\alias{test_lData}
\title{test_lData}
\usage{
test_lData(lData, lRef)
}
\arguments{
\item{lData}{a list with dataset}

\item{lRef}{a list with reference data}
}
\description{
Test raw data against reference data
}
