% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utils.R
\name{test_se_normalized}
\alias{test_se_normalized}
\title{test_se_normalized}
\usage{
test_se_normalized(se, lRef)
}
\arguments{
\item{se}{a SummarizedExperiment object with the Normalized assay}

\item{lRef}{a list of reference datasets}
}
\description{
Compare calculated normalization with the reference
}
