% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.utils.R
\name{test_synthetic_normalization}
\alias{test_synthetic_normalization}
\title{test_synthetic_normalization}
\usage{
test_synthetic_normalization(se, refNormalizedTsv)
}
\arguments{
\item{se}{a SummarizedExperiment object with the calculated assay}

\item{refNormalizedTsv}{a tsv file with the reference normalized synthetic data}
}
\description{
Compare calculated normalization with the reference in synthetic data
}
