% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_these_fxns_to_gDRwrapper.R
\name{update_experiment_metadata}
\alias{update_experiment_metadata}
\title{update_metadata}
\usage{
update_experiment_metadata(metadata, metadataList = NULL)
}
\arguments{
\item{SE}{a SummarizedExperiment}
}
\value{
a SummarizedExperiment with additional metadata
}
\description{
Update metadata in a SummarizedExperiment
}
