% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{update_metadata}
\usage{
update_metadata(
  SE,
  expert_unixid = Sys.getenv("USER"),
  description = NA,
  assay_id = NA,
  date_processed = NA,
  source_id = NA,
  state_id = NA,
  experiment_name = NA,
  qcs_id = NA,
  labhead_unixid = NA
)
}
\arguments{
\item{SE}{a SummarizedExperiment}
}
\value{
a SummarizedExperiment with additional metadata
}
\description{
Update metadata in a SummarizedExperiment
}
