% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_report.R
\name{validate_and_prepare_inputs}
\alias{validate_and_prepare_inputs}
\title{Validate and prepare script inputs based on data type}
\usage{
validate_and_prepare_inputs(args, data_type)
}
\arguments{
\item{args}{A list of all arguments passed to the main function (with staged paths).}

\item{data_type}{A character string of the data type, as determined by \code{determine_data_type}.}
}
\value{
A list of processed and validated arguments.
}
\description{
Performs detailed validation and preprocessing of arguments for the specified data type.
This includes checking the contents of annotation files.
}
\keyword{internal}
