% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{add_CellLine_annotation}
\alias{add_CellLine_annotation}
\title{add_CellLine_annotation}
\usage{
add_CellLine_annotation(df_metadata)
}
\arguments{
\item{df_metadata}{a dataframe with metadata}
}
\value{
a dataframe with metadata with annotated cell lines
}
\description{
add cellline annotation to a dataframe with metadata
}
\examples{
add_CellLine_annotation(df_metadata)
}
