% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_annotation.R
\name{add_Drug_annotation}
\alias{add_Drug_annotation}
\title{add_Drug_annotation}
\usage{
add_Drug_annotation(df_metadata, fill_DB_with_unknown = FALSE)
}
\arguments{
\item{df_metadata}{a data.frame with metadata}

\item{fill_DB_with_unknown}{a logical indicating whether DB should be filled with unknown drugs}
}
\value{
a data.frame with metadata with annotated drugs
}
\description{
add drug annotation to a data.frame with metadata
}
