% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_SE.R
\name{average_SE}
\alias{average_SE}
\title{average_SE}
\usage{
average_SE(normSE, TrtKeys = NULL, include_masked = F)
}
\arguments{
\item{normSE}{a SummarizedExperiment with normalized DR data}

\item{TrtKeys}{a vector of keys used for averaging (NULL by default)}

\item{include_masked}{boolean indicating whether or not to include the masked wells
in the averaging. 
Defaults to \code{FALSE}.}
}
\value{
a SummarizedExperiment with additional assay with averaged DR data
}
\description{
Avereage normalized SummarizedExperiment of DR data
}
