% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_SE.R
\name{average_SE2}
\alias{average_SE2}
\title{average_SE2}
\usage{
average_SE2(se, include_masked = FALSE, normalized_assay = "Normalized")
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} with drug response data.}

\item{trt_keys}{a vector of keys used for averaging (NULL by default)}
}
\value{
a SummarizedExperiment with additional assay with averaged DR data
}
\description{
Average the normalized data per
}
\seealso{
runDrugResponseProcessingPipeline2
}
