% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_SE.R
\name{average_SE2}
\alias{average_SE2}
\title{average_SE2}
\usage{
average_SE2(
  se,
  include_masked = FALSE,
  normalized_assay = "Normalized",
  averaged_assay = "Averaged"
)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} with drug response data that has a 
\code{normalized_assay}.}

\item{include_masked}{boolean indicating whether or not to include masked wells
in the averaging. 
This is used as an override to whatever wells have been masked in the original data.
Defaults to \code{FALSE}.}

\item{normalized_assay}{string of the assay name containing the normalized data.
Defaults to \code{"Normalized"}.}

\item{averaged_assay}{string of the assay to output averaged values in.
Defaults to \code{"Averaged"}.}
}
\value{
a \linkS4class{SummarizedExperiment} with an additional assay 
specified by \code{averaged_assay} containing normalized data with 
mean and standard deviation calculations for each unique treatment in the nested
\code{DataFrame}s.
}
\description{
Average the normalized data within the nested \code{DataFrame}s.
}
\seealso{
Other runDrugResponseProcessingPipelineFxns: 
\code{\link{create_SE2}()},
\code{\link{fit_SE2}()},
\code{\link{normalize_SE2}()},
\code{\link{runDrugResponseProcessingPipeline2}()}
}
\concept{runDrugResponseProcessingPipelineFxns}
