% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_SE.R
\name{average_SE2}
\alias{average_SE2}
\title{average_SE2}
\usage{
average_SE2(
  se,
  override_masked = FALSE,
  normalized_assay = "Normalized",
  averaged_assay = "Averaged"
)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} with drug response data that has a 
\code{normalized_assay}.}

\item{override_masked}{boolean indicating whether or not to override the masked wells
in the averaging and include all wells. 
Defaults to \code{FALSE}.}

\item{normalized_assay}{string of the assay name containing the normalized data.
Defaults to \code{"Normalized"}.}

\item{averaged_assay}{string of the assay to output averaged values in.
Defaults to \code{"Averaged"}.}
}
\value{
a \linkS4class{SummarizedExperiment} with an additional assay 
specified by \code{averaged_assay} containing normalized data with 
mean and standard deviation calculations for each unique treatment in the nested
\code{DataFrame}s.
}
\description{
Average the normalized data within the nested \code{DataFrame}s.
}
\details{
Expects that \code{gDRutils::get_SE_keys(se)} will have values for both 
\code{"Trt"} and \code{"masked_tag"}.
}
\seealso{
Other runDrugResponseProcessingPipelineFxns: 
\code{\link{create_SE2}()},
\code{\link{fit_SE2}()},
\code{\link{normalize_SE2}()},
\code{\link{runDrugResponseProcessingPipeline2}()}
}
\concept{runDrugResponseProcessingPipelineFxns}
