% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_GR_value.R
\name{calculate_GR_value}
\alias{calculate_GR_value}
\title{Calculate a GR value.}
\usage{
calculate_GR_value(df_, cellline_md, day0_readout, ndigit_rounding)
}
\arguments{
\item{df_}{data.frame of dose response values containing \code{"CorrectedReadout"}.}

\item{cellline_md}{named list of metadata for the cell line of interest.}

\item{day0_readout}{numeric vector}

\item{ndigit_rounding}{integer}
}
\value{

}
\description{
Calculate a GR value for a given set of dose response values.
}
