% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{cleanup_metadata}
\alias{cleanup_metadata}
\title{cleanup_metadata}
\usage{
cleanup_metadata(df_metadata)
}
\arguments{
\item{df_metadata}{a dataframe with metadata}
}
\value{
a dataframe with cleaned metadata
}
\description{
Cleanup a dataframe with metadata
}
\examples{
\dontrun{
cleanup_metadata(df_metadata)
}
}
