% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_utils.R
\name{create_control_df}
\alias{create_control_df}
\title{Create a control dataframe for a treatment-cell line combination.}
\usage{
create_control_df(df_, Keys, key, control_mean_fxn, out_col_name)
}
\arguments{
\item{out_col_name}{string of the output readout that will replace \code{CorrectedReadout}.}
}
\description{
Create a control dataframe for a treatment-cell line combination.
}
