% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SE.R
\name{create_SE2}
\alias{create_SE2}
\title{Create a SummarizedExperiment object}
\usage{
create_SE2(
  df_,
  readout = "ReadoutValue",
  control_mean_fxn = function(x) {     mean(x, trim = 0.25) },
  discard_keys = c("Barcode", gDRutils::get_identifier("masked_tag"))
)
}
\arguments{
\item{df_}{data.frame of raw drug response data containing both treated and untreated values.}

\item{readout}{string of the name containing the cell viability readout values.}

\item{control_mean_fxn}{function indicating how to average controls.
Defaults to \code{mean(x, trim = 0.25)}.}

\item{discard_keys}{character vector of column names to include in the data.frames in the assays of the resulting \code{SummarizedExperiment} object. 
Defaults to \code{NULL}.}
}
\value{
A \linkS4class{SummarizedExperiment} object containing two asssays.
Treated readouts will live in an assay called \code{"RawTreated"},
and reference readouts live in an assay called \code{"Controls"}.
\code{rownames} and \code{colnames} are made up of available metadata on treatments and conditions 
and is pasted together.
}
\description{
Create a SummarizedExperiment object from a data.frame where the data.frame contains treatments on rows,
conditions are on columns.
}
\details{
This is most commonly used in preparation for downstream normalization.
}
\seealso{
normalize_SE2, runDrugResponseProcessingPipeline
}
