% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SE.R
\name{create_SE2}
\alias{create_SE2}
\title{Create a SummarizedExperiment object}
\usage{
create_SE2(df_, readout = "ReadoutValue", discard_keys = NULL)
}
\arguments{
\item{df_}{data.frame of raw drug response data containing both treated and untreated values.}

\item{readout}{string of the name containing the cell viability readout values.}

\item{discard_keys}{character vector of column names to include in the data.frames in the assays of the resulting \code{SummarizedExperiment} object. 
Defaults to \code{NULL}.}
}
\description{
Create a SummarizedExperiment object from a data.frame where all treatments are on rows,
conditions are on columns, treated readouts live in an assay called \code{"treated"},
and reference readouts live in an assay called \code{"Controls"}.
}
\details{
This is most commonly used in preparation for downstream normalization.
}
\seealso{
normalize_SE
}
