% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\name{createSE}
\alias{createSE}
\title{createSE}
\usage{
createSE(df_data, data_type = c("untreated", "treated", "all"),
  readout = "ReadoutValue", discard_keys = NULL)
}
\arguments{
\item{data_type}{string type of data to be returned: with untreated conditions only ('untreated'),
with treated conditions only ('treated') or all}

\item{dfList}{list with data.frame(s) with dose-reponse data}
}
\value{
SummarizedExperiment object with dose-reponse data
}
\description{
Create SummarizedExperiment object from data.frame(s) with dose-reponse data
}
