% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_treated_and_untreated_conditions.R
\name{.assign_treated_and_untreated_conditions}
\alias{.assign_treated_and_untreated_conditions}
\title{.assign_treated_and_untreated_conditions}
\usage{
.assign_treated_and_untreated_conditions(df_)
}
\arguments{
\item{df_}{containing the drug name field as specified by \code{gDRutils::get_identifier("drugname")}..}
}
\value{
data.frame containing an additional factor column called \code{treated_untreated} specifying
whether the condition is a treated or untreated entry.
}
\description{
Assign the treated and untreated conditions to a data.frame
}
