% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treated_conditions.R
\name{.get_untreated_conditions}
\alias{.get_untreated_conditions}
\title{.get_untreated_conditions}
\usage{
.get_untreated_conditions(drug_data)
}
\arguments{
\item{drug_data}{data.frame or DataFrame with treatment information}
}
\value{
character vector with untreated conditions
}
\description{
Get untreated conditions
}
