% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_SE.R
\name{fit_SE2}
\alias{fit_SE2}
\title{fit_SE2}
\usage{
fit_SE2(
  se,
  averaged_assay = "Averaged",
  metrics_assay = "Metrics",
  studyConcThresh = 4
)
}
\arguments{
\item{se}{a \linkS4class{SummarizedExperiment} with averaged and normalized assays}

\item{averaged_assay}{string of the name of the averaged assay in the \linkS4class{SummarizedExperiment}.}

\item{metrics_assay}{string of the metrics assay to output.}

\item{studyConcThresh}{a numeric with study concentration threshold (4 by default)}
}
\value{
the original \linkS4class{SummarizedExperiment} with an additional assay named \code{metrics_assay}
}
\description{
Calculate metrics for DR data
}
