% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gDR-package}
\alias{gDR}
\alias{gDR-package}
\title{gDR: Umbrella package for R packages in the gDR suite}
\description{
Package is a part of the gDR suite. It provides info about processing functions and utilities from individual gDR packages. The vignette walks through the full processing pipeline for drug response analyses that the gDR suite offers.
}
\author{
\strong{Maintainer}: Support gDR \email{gdr-support-d@gene.com} (\href{https://orcid.org/0009-0006-9310-1135}{ORCID})

Authors:
\itemize{
  \item Allison Vuong \email{vuong.allison@gene.com}
  \item Bartosz Czech
  \item Arkadiusz Gladki
  \item Marc Hafner
  \item Dariusz Scigocki
  \item Sergiu Mocanu
}

}
\keyword{internal}
