% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDR-package.R, R/package.R
\docType{package}
\name{gDR-package}
\alias{gDR-package}
\title{gDR: Umbrella package for R packages in the gDR suite}
\value{
package help page

package help page
}
\description{
Package is a part of the gDR suite. It reexports functions from other packages in the gDR suite that contain critical processing functions and utilities. The vignette walks through the full processing pipeline for drug response analyses that the gDR suite offers.

Package is a part of the gDR suite. It reexports functions from other packages in the gDR suite that contain critical processing functions and utilities. The vignette walks through the full processing pipeline for drug response analyses that the gDR suite offers.
}
\note{
To learn more about functions start with \code{help(package = "gDR")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gdrplatform/gDR}
  \item \url{https://gdrplatform.github.io/gDR/}
  \item Report bugs at \url{https://github.com/gdrplatform/gDR/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/gdrplatform/gDR}
  \item \url{https://gdrplatform.github.io/gDR/}
  \item Report bugs at \url{https://github.com/gdrplatform/gDR/issues}
}

}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})

Authors:
\itemize{
  \item Allison Vuong \email{vuong.allison@gene.com}
  \item Bartosz Czech (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
  \item Marc Hafner (\href{https://orcid.org/0000-0003-1337-7598}{ORCID})
  \item Dariusz Scigocki
  \item Janina Smola
  \item Sergiu Mocanu
}

}
\keyword{internal}
