% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R, R/getMetaData.R
\name{getMetaData}
\alias{getMetaData}
\title{getMetaData}
\usage{
getMetaData(
  data,
  cell_id = gDRutils::get_identifier("cellline"),
  discard_keys = NULL
)

getMetaData(
  data,
  cell_id = gDRutils::get_identifier("cellline"),
  discard_keys = NULL
)
}
\arguments{
\item{data}{tibble or data.frame with drug-response data}

\item{cell_id}{string name of the column with cell line names}
}
\value{
list with two DataFrames ('colData' and 'rowData') and character vector ('dataCols')

list with two DataFrames ('colData' and 'rowData') and character vector ('dataCols')
}
\description{
Get metadata of given project

Get metadata of given project
}
