% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_curve_fit.R
\name{ICGRfits}
\alias{ICGRfits}
\title{Actual fitting function}
\usage{
ICGRfits(df_, e_0 = 1, GR_0 = 1, force = FALSE, cap = FALSE)
}
\arguments{
\item{e_0}{=1 by default}

\item{GR_0}{=1 by default}

\item{force}{use signifcance or not}

\item{cap}{enforce e_0 and GR_0}

\item{log10concs}{concentrations}

\item{RelativeViability}{values}

\item{GRvalues}{values}
}
\value{
vector of parameters
}
\description{
\code{ICGRfits} returns fit parameters
}
\details{
returns fit parameters
}
\examples{
sum(1:10)
}
