% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data.R
\name{identify_keys}
\alias{identify_keys}
\title{identify_keys}
\usage{
identify_keys(df_se_mae)
}
\arguments{
\item{df_se_mae}{a dataframe or SummarizedExperiment or MultiassayExperiment with keys}
}
\value{
a list of keys
}
\description{
Identify keys in the DR data represented by dataframe or SummarizedExperiment or MultiAssayExperiment objects
}
