% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import raw data}
\usage{
import_data(
  manifest_file,
  template_file,
  results_file,
  instrument = "EnVision"
)
}
\arguments{
\item{manifest_file}{character, file path(s) to manifest(s)}

\item{template_file}{data.frame, with datapaths and names of results file(s)
or character with file path of templates file(s)}

\item{results_file}{data.frame, with datapaths and names of results file(s)
or character with file path of results file(s)}

\item{instrument}{character with type of instrument used}
}
\value{
a \code{data.frame}
}
\description{
Import raw data
}
\examples{
td <- gDRimport::get_test_data()
i_df <- import_data(td$m_file, td$t_files, td$r_files)

}
