% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Order_result_df}
\alias{Order_result_df}
\title{Order_result_df}
\usage{
Order_result_df(df_)
}
\arguments{
\item{df_}{a dataframe with results}
}
\value{
a ordered dataframe with results
}
\description{
Order a dataframe with results
}
