% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRcore.R, R/gDRimport.R, R/gDRutils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{merge_data}
\alias{runDrugResponseProcessingPipeline}
\alias{create_SE}
\alias{normalize_SE}
\alias{average_SE}
\alias{fit_SE}
\alias{load_data}
\alias{fit_curves}
\alias{logisticFit}
\alias{get_header}
\alias{reset_headers}
\alias{get_identifier}
\alias{set_identifier}
\alias{reset_identifiers}
\alias{convert_se_assay_to_dt}
\alias{validate_SE}
\alias{flatten}
\alias{split_SE_components}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gDRcore}{\code{\link[gDRcore]{average_SE}}, \code{\link[gDRcore]{create_SE}}, \code{\link[gDRcore]{fit_SE}}, \code{\link[gDRcore]{merge_data}}, \code{\link[gDRcore]{normalize_SE}}, \code{\link[gDRcore]{runDrugResponseProcessingPipeline}}}

  \item{gDRimport}{\code{\link[gDRimport]{load_data}}}

  \item{gDRutils}{\code{\link[gDRutils]{convert_se_assay_to_dt}}, \code{\link[gDRutils]{fit_curves}}, \code{\link[gDRutils]{flatten}}, \code{\link[gDRutils:headers]{get_header}}, \code{\link[gDRutils:identifiers]{get_identifier}}, \code{\link[gDRutils]{logisticFit}}, \code{\link[gDRutils:headers]{reset_headers}}, \code{\link[gDRutils:identifiers]{reset_identifiers}}, \code{\link[gDRutils:identifiers]{set_identifier}}, \code{\link[gDRutils]{split_SE_components}}, \code{\link[gDRutils]{validate_SE}}}
}}

