% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRcomponents.R, R/gDRcore.R, R/gDRimport.R,
%   R/gDRutils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{save_config_template}
\alias{load_config}
\alias{add_project_to_metaproject}
\alias{add_metaproject_to_metaprojects_list}
\alias{add_project_to_projects_list}
\alias{save_MAE}
\alias{getProjects}
\alias{getMetaProjects}
\alias{pullProjectSE}
\alias{merge_data}
\alias{runDrugResponseProcessingPipeline}
\alias{create_and_normalize_SE}
\alias{normalize_SE}
\alias{average_SE}
\alias{fit_SE}
\alias{load_data}
\alias{fit_curves}
\alias{logisticFit}
\alias{get_header}
\alias{get_env_identifiers}
\alias{get_SE_identifiers}
\alias{set_env_identifier}
\alias{set_SE_identifiers}
\alias{reset_env_identifiers}
\alias{convert_se_assay_to_dt}
\alias{validate_SE}
\alias{flatten}
\alias{split_SE_components}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gDRcomponents}{\code{\link[gDRcomponents]{add_metaproject_to_metaprojects_list}}, \code{\link[gDRcomponents]{add_project_to_metaproject}}, \code{\link[gDRcomponents]{add_project_to_projects_list}}, \code{\link[gDRcomponents]{getMetaProjects}}, \code{\link[gDRcomponents]{getProjects}}, \code{\link[gDRcomponents]{load_config}}, \code{\link[gDRcomponents]{pullProjectSE}}, \code{\link[gDRcomponents]{save_config_template}}, \code{\link[gDRcomponents]{save_MAE}}}

  \item{gDRcore}{\code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{average_SE}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{create_and_normalize_SE}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{fit_SE}}, \code{\link[gDRcore]{merge_data}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{normalize_SE}}, \code{\link[gDRcore:runDrugResponseProcessingPipelineFxns]{runDrugResponseProcessingPipeline}}}

  \item{gDRimport}{\code{\link[gDRimport]{load_data}}}

  \item{gDRutils}{\code{\link[gDRutils]{convert_se_assay_to_dt}}, \code{\link[gDRutils]{fit_curves}}, \code{\link[gDRutils]{flatten}}, \code{\link[gDRutils:identifiers]{get_env_identifiers}}, \code{\link[gDRutils:headers]{get_header}}, \code{\link[gDRutils:SE_metadata]{get_SE_identifiers}}, \code{\link[gDRutils]{logisticFit}}, \code{\link[gDRutils:identifiers]{reset_env_identifiers}}, \code{\link[gDRutils:identifiers]{set_env_identifier}}, \code{\link[gDRutils:SE_metadata]{set_SE_identifiers}}, \code{\link[gDRutils]{split_SE_components}}, \code{\link[gDRutils]{validate_SE}}}
}}

